pyproject.toml
[tool.poetry]
name = "builds"
version = "0.1.0"
description = ""
authors = ["gevlen <gevlenko1993@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
orjson = "^3.8.9"
httpx = "^0.23.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
mypy = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
explicit_package_bases = true